@ndgrigorian 2023-09-03
@oleksandr-pavlyk 
I wasn't sure if you were planning to add more to this PR.

If not, I'll go ahead and do a last look-over and approve it. I tested it out last week and found no correctness issues and the performance is much better.